backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
beanie
fastapi-users[beanie]